Fixes for x64 Windows builds.

+++ src/maidsafe/encrypt/tests/self_encryptor_test.cc
@@ -1182,16 +1182,16 @@ TEST_F(BasicTest, BEH_ManualCheckWrite) {
 }
 
 TEST_F(BasicTest, BEH_TruncateIncreaseScenario1) {
-  const size_t kTestDataSize(kDefaultChunkSize * 12);
+  const uint32_t kTestDataSize(kDefaultChunkSize * 12);
   boost::scoped_array<char>plain_data(new char[kTestDataSize]);
   memset(plain_data.get(), 0, kTestDataSize);
 
-  for (size_t i = 0; i < kTestDataSize; i += 4096) {
+  for (uint32_t i = 0; i < kTestDataSize; i += 4096) {
     uint32_t write_position(i);
     uint32_t write_length(4096);
     std::string plain_text(RandomString(write_length));
     boost::scoped_array<char>content_data(new char[write_length]);
-    for (size_t i = 0; i < write_length; ++i) {
+    for (uint32_t i = 0; i < write_length; ++i) {
       plain_data[i + write_position] = plain_text[i];
       content_data[i] = plain_text[i];
     }
@@ -1207,7 +1207,7 @@ TEST_F(BasicTest, BEH_TruncateIncreaseScenario1) {
   boost::scoped_array<char>answer(new char[read_length]);
   EXPECT_TRUE(self_encryptor_->Read(answer.get(),
                                     read_length, read_position));
-  for (size_t i = 0; i < read_length; ++i)
+  for (uint32_t i = 0; i < read_length; ++i)
     if ((i + read_position) < self_encryptor_->size())
       ASSERT_EQ(plain_data[read_position + i], answer[i])
           << "not match " << i << " from " << read_position
